# docs/swagger.yaml
basePath: /
definitions:
  dto.BCVResponse:
    properties:
      eur:
        example: 39.1234
        type: number
      fecha:
        example: "2026-03-20T00:00:00Z"
        type: string
      usd:
        example: 36.4521
        type: number
    type: object
  dto.CiudadResponse:
    properties:
      estado_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      nombre:
        example: Barquisimeto
        type: string
    type: object
  dto.CreateUCAUURequest:
    properties:
      valor:
        description: validate:"required,gt=0" asegura que el campo venga y sea mayor
          a cero
        example: 35.5
        type: number
    required:
    - valor
    type: object
  dto.EstadoResponse:
    properties:
      id:
        example: 1
        type: integer
      nombre:
        example: Lara
        type: string
    type: object
  dto.GenericResponse:
    properties:
      data: {}
      message:
        example: Operación exitosa
        type: string
    type: object
  dto.MunicipioResponse:
    properties:
      estado_id:
        example: 1
        type: integer
      id:
        example: 10
        type: integer
      nombre:
        example: Iribarren
        type: string
    type: object
  dto.ParroquiaResponse:
    properties:
      id:
        example: 50
        type: integer
      municipio_id:
        example: 10
        type: integer
      nombre:
        example: Catedral
        type: string
    type: object
  dto.TribunalResponse:
    properties:
      categoria:
        example: Superior
        type: string
      id:
        example: 1
        type: integer
      nombre:
        example: Corte de Apelaciones Sala 1
        type: string
    type: object
  dto.UCAUUResponse:
    properties:
      fecha:
        example: "2026-03-20T10:00:00Z"
        type: string
      valor:
        example: 35.5
        type: number
    type: object
info:
  contact:
    name: Jose
  description: API centralizada para indicadores económicos y territoriales.
  title: API Global Universitas
  version: "1.0"
paths:
  /api/v1/economia/bcv:
    get:
      description: Retorna el valor actual del Dólar y Euro oficial.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.GenericResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BCVResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener tasas del BCV (Dólar y Euro)
      tags:
      - Economia
  /api/v1/economia/ucauu:
    get:
      description: Retorna el último valor registrado de la unidad UCAUU
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UCAUUResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener última UCAUU
      tags:
      - Economia
    post:
      consumes:
      - application/json
      description: Permite registrar el valor de la UCAUU del día actual. Si ya existe,
        lo actualiza.
      parameters:
      - description: Valor de la UCAUU
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUCAUURequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Registrar o Actualizar UCAUU
      tags:
      - Economia
  /api/v1/territorio/estados:
    get:
      description: Retorna la lista de todos los estados de Venezuela
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.GenericResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.EstadoResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener todos los estados
      tags:
      - Territorio
  /api/v1/territorio/estados/{estado_id}/ciudades:
    get:
      description: Retorna la lista de ciudades pertenecientes a un estado específico
      parameters:
      - description: ID del Estado
        in: path
        name: estado_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.GenericResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CiudadResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener ciudades por estado
      tags:
      - Territorio
  /api/v1/territorio/estados/{estado_id}/municipios:
    get:
      description: Retorna la lista de municipios pertenecientes a un estado específico
      parameters:
      - description: ID del Estado
        in: path
        name: estado_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.GenericResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MunicipioResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener municipios por estado
      tags:
      - Territorio
  /api/v1/territorio/estados/{estado_id}/tribunales:
    get:
      description: Retorna la lista de tribunales de categoría Superior y Primera
        Instancia pertenecientes a un estado
      parameters:
      - description: ID del Estado
        in: path
        name: estado_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.GenericResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TribunalResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener tribunales estadales por estado
      tags:
      - Territorio
  /api/v1/territorio/municipios/{municipio_id}/parroquias:
    get:
      description: Retorna la lista de parroquias pertenecientes a un municipio específico
      parameters:
      - description: ID del Municipio
        in: path
        name: municipio_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.GenericResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ParroquiaResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener parroquias por municipio
      tags:
      - Territorio
  /api/v1/territorio/municipios/{municipio_id}/tribunales:
    get:
      description: Retorna la lista de tribunales de categoría Municipio asociados
        a un municipio específico, incluyendo los compartidos con otros municipios
      parameters:
      - description: ID del Municipio
        in: path
        name: municipio_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.GenericResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TribunalResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse'
      summary: Obtener tribunales municipales por municipio
      tags:
      - Territorio
swagger: "2.0"